take 1
bench: 2406216

+++ src/transposition_table.hpp
@@ -15,14 +15,14 @@ enum class TFlag: uint8_t {
 
 // side to move is not stored in the transposition table as it is in the zobrist hash
 struct TEntry {
-    uint64_t zobrist_hash   = 0; // 8 bytes
+    uint32_t zobrist_hash   = 0; // 4 bytes
     int16_t value           = NO_VALUE; // 2 bytes
     int16_t static_eval     = NO_VALUE; // 2 bytes
     uint16_t move           = 0; // 2 bytes
     uint8_t depth_tflag     = 0; // 1 byte  -> contains depth: 6 bits (64 values), flag: 2 bits (4 values)
     uint8_t move_number     = 0; // 1 byte
     // ==============
-    // ----> total = 8 + 8 = 16 bytes
+    // ----> total = 4 + 8 = 12 bytes
 
     int depth(){
         return static_cast<int>(depth_tflag >> 2);
@@ -34,7 +34,7 @@ struct TEntry {
 
     TEntry(){};
 
-    TEntry(uint64_t zobrist, int value, int static_eval, int depth, Move move, TFlag flag, uint8_t move_number):
+    TEntry(uint32_t zobrist, int value, int static_eval, int depth, Move move, TFlag flag, uint8_t move_number):
             zobrist_hash(zobrist),
             value(value),
             static_eval(static_eval),
@@ -44,7 +44,7 @@ struct TEntry {
 };
 
 struct TTData {
-    uint64_t zobrist_hash   = 0;
+    uint32_t zobrist_hash   = 0;
     int value               = NO_VALUE;
     int static_eval         = NO_VALUE;
     Move move               = Move::NO_MOVE;
@@ -54,7 +54,7 @@ struct TTData {
 
     TTData(){};
 
-    TTData(uint64_t zobrist, int value, int static_eval, int depth, Move move, TFlag flag, uint8_t move_number):
+    TTData(uint32_t zobrist, int value, int static_eval, int depth, Move move, TFlag flag, uint8_t move_number):
             zobrist_hash(zobrist),
             value(value),
             static_eval(static_eval),
@@ -80,9 +80,9 @@ class TranspositionTable {
 
     void allocateMB(int new_size);
 
-    void store(uint64_t zobrist, int value, int eval, int depth, Move move, TFlag flag, uint8_t move_number);
+    void store(uint32_t zobrist, int value, int eval, int depth, Move move, TFlag flag, uint8_t move_number);
 
-    TTData probe(bool& is_hit, uint64_t zobrist);
+    TTData probe(bool& is_hit, uint32_t zobrist);
 
     void clear();
